src/main/webapp/app/entities/materiel/delete/materiel-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { MaterielService } from '../service/materiel.service';

import { MaterielDeleteDialogComponent } from './materiel-delete-dialog.component';

describe('Component Tests', () => {
  describe('Materiel Management Delete Component', () => {
    let comp: MaterielDeleteDialogComponent;
    let fixture: ComponentFixture<MaterielDeleteDialogComponent>;
    let service: MaterielService;
    let mockActiveModal: NgbActiveModal;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [MaterielDeleteDialogComponent],
        providers: [NgbActiveModal],
      })
        .overrideTemplate(MaterielDeleteDialogComponent, '')
        .compileComponents();
      fixture = TestBed.createComponent(MaterielDeleteDialogComponent);
      comp = fixture.componentInstance;
      service = TestBed.inject(MaterielService);
      mockActiveModal = TestBed.inject(NgbActiveModal);
    });

    describe('confirmDelete', () => {
      it('Should call delete service on confirmDelete', inject(
        [],
        fakeAsync(() => {
          // GIVEN
          spyOn(service, 'delete').and.returnValue(of({}));

          // WHEN
          comp.confirmDelete(123);
          tick();

          // THEN
          expect(service.delete).toHaveBeenCalledWith(123);
          expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
        })
      ));

      it('Should not call delete service on clear', () => {
        // GIVEN
        spyOn(service, 'delete');

        // WHEN
        comp.cancel();

        // THEN
        expect(service.delete).not.toHaveBeenCalled();
        expect(mockActiveModal.close).not.toHaveBeenCalled();
        expect(mockActiveModal.dismiss).toHaveBeenCalled();
      });
    });
  });
});
